requirements.txt
replicate
Pillow
python-dotenv
streamlit